+++ modules/client/src/internal/client/methods.ts
@@ -85,6 +85,7 @@ import {
   LIVE_CONTRACTS,
   MultiTargetPermission,
   SortDirection,
+  SupportedNetworks,
   TokenType,
 } from "../../client-common";
 import {
