+++ modules/client/src/tokenVoting/internal/client/estimation.ts
@@ -8,10 +8,7 @@ import {
   GasFeeEstimation,
   VoteProposalParams,
 } from "../../../client-common";
-import {
-  DelegateTokensParams,
-  ITokenVotingClientEstimation,
-} from "../../interfaces";
+import { ITokenVotingClientEstimation } from "../../interfaces";
 import { toUtf8Bytes } from "@ethersproject/strings";
 import {
   boolArrayToBitmap,
@@ -19,6 +16,7 @@ import {
   NoProviderError,
   NoSignerError,
 } from "@aragon/sdk-common";
+import { DelegateTokensParams } from "../../types";
 /**
  * Estimation module the SDK TokenVoting Client
  */
@@ -135,7 +133,7 @@ export class TokenVotingClientEstimation extends ClientCore
     );
     return this.web3.getApproximateGasFee(estimation.toBigInt());
   }
-  
+
   /**
    * Estimates the gas fee of delegating voting power to a delegatee
    *
