+++ modules/client/src/tokenVoting/internal/types.ts
@@ -71,3 +71,16 @@ export type SubgraphTokenVotingProposal =
     executionTxHash: string;
     executionBlockNumber: string;
   };
+
+  export type SubgraphTokenVotingMember = {
+    address: string;
+    balance: string;
+    votingPower: string;
+    delegatee: {
+      address: string;
+    };
+    delegators: {
+      address: string;
+      balance: string;
+    }[];
+  };
