+++ modules/client/test/integration/client/encoding.test.ts
@@ -10,6 +10,7 @@ import {
 
 import {
   ApplyInstallationParams,
+  ApplyUninstallationParams,
   Client,
   Context,
   DaoMetadata,
