+++ modules/client/test/integration/client/methods.test.ts
@@ -31,9 +31,11 @@ import {
   DaoSortBy,
   DepositParams,
   HasPermissionParams,
+  LIVE_CONTRACTS,
   Permissions,
   PluginQueryParams,
   PluginSortBy,
+  PrepareUninstallationSteps,
   SetAllowanceParams,
   SetAllowanceSteps,
   SortDirection,
