+++ modules/client/test/integration/tokenVoting-client/methods.test.ts
@@ -8,6 +8,7 @@ import {
   Context,
   ContextPlugin,
   CreateMajorityVotingProposalParams,
+  DelegateTokensStep,
   Erc20TokenDetails,
   ExecuteProposalStep,
   PrepareInstallationStep,
@@ -22,8 +23,10 @@ import {
   SupportedNetworksArray,
   TokenType,
   TokenVotingClient,
+  TokenVotingMember,
   TokenVotingPluginPrepareInstallationParams,
   TokenVotingProposal,
+  UndelegateTokensStep,
   UnwrapTokensStep,
   VoteProposalParams,
   VoteProposalStep,
